WIP: mount compatibility on darwin
CoreOS disallows mounting at `/`root because it is
[immutable](https://docs.fedoraproject.org/en-US/fedora-coreos/storage/#_immutable_read_only_usr).

As an alternative, it is recommended to mount at `/mnt`, e.g:
`podman machine init -v /Users:/mnt/Users` on darwin. This, however,
breaks compatibility with other tools, e.g. docker desktop, colima when
specifying mounts:
```config
\# docker-compose.yml
...
volumes:
    - ./app:/app
```
- Works in colima and docker desktop.
- Fails with `podman run` because it is missing the `/mnt` prefix.

As a result, I can't use podman and docker desktop interchangeably
with one single `docker-compose.yml` file.

This WIP tries to solve that with the following convention:
- `podman machine:` every target mount path into coreOS is internally prefixed with
`/mnt` so that it can be mounted.
- `podman run`: every source mount path is internally prefixed with `/mnt` if you
are on darwin.

Signed-off-by: Thibault Gagnaux <tgagnaux@gmail.com>